split generator
