+++ generators/entities/index.mjs
@@ -1,0 +1,1 @@
+export { default } from './generator.mjs';
